Remove warnings
git-svn-id: file:///.data/svn_convert/svn@276 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/client.cpp
@@ -864,7 +864,7 @@ void QClientSocket::write(const char *buf, unsigned int size)
     bInWrite = true;
     int res = sock->writeBlock(buf, size);
     bInWrite = false;
-    if (res != size){
+    if (res != (int)size){
         log(L_WARN, "Write error");
         notify->error_state();
         return;
